# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Directorio de empresas - CCGAMT",
        "description": "Backend y API Acceso a datos.",
        "contact": {
            "email": "brayan-diaz3@upc.edu.co"
        },
        "license": {
            "name": "@Universidad Piloto de Colombia SAM - Todos los Derechos Reservados"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/comentarios": {
            "get": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Listar todos los comentarios",
                "operationId": "8478472babe96435537abd02f58aee0d",
                "responses": {
                    "200": {
                        "description": "Lista de comentarios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Crear un nuevo comentario",
                "operationId": "5da383245219a9b40dbe89d787ea634c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "foro_id",
                                    "contenido",
                                    "fecha_publicacion"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "foro_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "contenido": {
                                        "type": "string",
                                        "example": "Este es un comentario de ejemplo."
                                    },
                                    "fecha_publicacion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comentario creado correctamente"
                    },
                    "400": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/comentarios/{id}": {
            "get": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Obtener un comentario por ID",
                "operationId": "69ee0321df0feb425bd2be40835c5f2b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comentario encontrado"
                    },
                    "404": {
                        "description": "Comentario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Actualizar un comentario",
                "operationId": "fc3e9975be76ff9587fd4c04d0ecc1df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "contenido",
                                    "fecha_publicacion"
                                ],
                                "properties": {
                                    "contenido": {
                                        "type": "string",
                                        "example": "Este es un comentario actualizado."
                                    },
                                    "fecha_publicacion": {
                                        "type": "string",
                                        "example": "2025-04-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comentario actualizado correctamente"
                    },
                    "400": {
                        "description": "Errores de validación"
                    },
                    "404": {
                        "description": "Comentario no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Eliminar un comentario",
                "operationId": "1e0b0839c86abd2906a32a05638a79d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comentario eliminado correctamente"
                    },
                    "404": {
                        "description": "Comentario no encontrado"
                    }
                }
            }
        },
...
    },
...
}